// webapp/src/app/core/api/v1/model/application.ts
/**
 * permaAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Application { 
    applicationId?: number;
    applicationTypeId?: number;
    applicationType?: string | null;
    memberId?: number;
    memberName?: string | null;
    applicationName?: string | null;
    readonly?: boolean;
    coverageYear?: number;
    currentSectionId?: number | null;
    applicationStatus?: string | null;
    dueDate?: string | null;
    completedDate?: string | null;
    completedBy?: string | null;
    previousApplicationId?: number | null;
}